+++ templates/dashboard.html
@@ -14,7 +14,7 @@
   <h2 class="mb-4">📋 메뉴 선택</h2>
   <div class="d-grid gap-3 w-100" style="max-width: 320px; margin: auto;">
     <a href="{{ url_for('explore') }}" class="btn btn-purple w-100">💘 내 이상형 찾기</a>
-    <a href="{{ url_for('liked_users') }}" class="btn btn-purple w-100">❤️ 내가 좋아요한 사람</a>
+    <a href="{{ url_for('liked_users.liked_users') }}" class="btn btn-purple w-100">❤️ 내가 좋아요한 사람</a>
     <a href="{{ url_for('matches') }}" class="btn btn-purple w-100">💑 매칭된 사람 보기</a>
     <a href="{{ url_for('chatlist') }}" class="btn btn-purple w-100"> ✉ 채팅 목록 </a>
     <a href="{{ url_for('profile') }}" class="btn btn-purple w-100">👤 내 프로필</a>
