+++ templates/profile.html
@@ -17,7 +17,7 @@
     <h2 class="fw-bold" style="font-size: 24px; margin: 0;">
       {{ user.nickname }}님의 프로필
     </h2>
-    <a href="{{ url_for('edit_profile') }}" class="btn btn-outline-secondary d-flex align-items-center" style="font-size: 14px; padding: 6px 12px;">
+    <a href="{{ url_for('profile.edit_profile') }}" class="btn btn-outline-secondary d-flex align-items-center" style="font-size: 14px; padding: 6px 12px;">
       ✏️ <span class="ms-1">수정</span>
     </a>
   </div>
